// src/app/api/generated/model/reporte-disponible.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SeccionDisponibleResponse } from './seccion-disponible-response';


/**
 * Un reporte y sus secciones
 */
export interface ReporteDisponible { 
    reporte?: string;
    secciones?: Array<SeccionDisponibleResponse>;
}